Begin to add styling
