+++ index.html
@@ -6,7 +6,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1">
 
   <!-- Do not add `link` tags-->
-  <link rel="shortcut icon" type="image/x-icon" href="favicon.ico">
+  <link rel="shortcut icon" type="image/x-icon" href="assets/styles/img/concha.png">
 
   <!-- Do not add `script` tags-->
   <script src="public/vendor.js" type="text/javascript" charset="utf-8" defer></script>
@@ -72,13 +72,14 @@
   </nav>
 
   <!-- end of nav -->
-  <h1>Tic Tac Toe: Pan Dulce Edition</h1>
+  <img src="assets/styles/img/picado.png" alt="Papel Picado">
+  <h1 class="title">Tic Tac Toe: Pan Dulce Edition</h1>
 
-  <div id="turn">
+  <div class="turn">
     <h2>Who's turn is it:</h2>
     <p id="turn-message"></p>
   </div>
-  <div id="won">
+  <div class="won">
     <h2>Who Won:</h2>
     <p id="won-message"></p>
   </div>
